requirements-test.txt
#! Test requirements for Smart Cover Automation

pytest>=8.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-homeassistant-custom-component>=0.13.0

# Static type checking
mypy>=1.0.0
